:truck: Divide keyboards to separate files.
