+++ services/keyboard/call.py
@@ -1,34 +1,9 @@
-from vkwave.bots.utils.keyboards import Keyboard
+from vkwave.bots import Keyboard
 
 from database import utils as db
+from services import keyboard as kbs
 
 JSONStr = str
-
-
-def main_menu(user_id: int) -> JSONStr:
-    """
-    Генерирует клавиатуру главного меню
-    Args:
-        user_id: Идентификатор пользователя
-    Returns:
-        JSONStr: Строка с клавиатурой
-
-    """
-    is_admin = db.admin.is_user_admin(
-        admin_id=db.students.get_system_id_of_student(user_id)
-    )
-    kb = Keyboard()
-    if is_admin:
-        kb.add_text_button(text="📢 Призыв", payload={"button": "call"})
-        kb.add_text_button(text="💰 Финансы", payload={"button": "finances"})
-        kb.add_row()
-    kb.add_text_button(text="📅 Расписание", payload={"button": "schedule"})
-    kb.add_text_button(text="📨 Рассылки", payload={"button": "mailings"})
-    if is_admin:
-        kb.add_row()
-        kb.add_text_button(text="⚙ Настройки", payload={"button": "settings"})
-        kb.add_text_button(text="🌐 Веб", payload={"button": "web"})
-    return kb.get_keyboard()
 
 
 def skip_call_message() -> JSONStr:
@@ -44,26 +19,6 @@ def skip_call_message() -> JSONStr:
     return kb.get_keyboard()
 
 
-def alphabet(user_id: int) -> Keyboard:
-    """
-    Генерирует фрагмент клавиатуры со списком первых букв фамилиий студентов
-
-    Args:
-        user_id: Идентификатор администратора
-
-    Returns:
-        Keyboard: Фрагмент клавиатуры
-    """
-    kb = Keyboard()
-    alphabet = db.students.get_unique_second_name_letters_in_a_group(user_id)
-    for letter in alphabet:
-        if len(kb.buttons[-1]) == 4:
-            kb.add_row()
-        kb.add_text_button(text=letter, payload={"button": "letter", "value": letter})
-
-    return kb
-
-
 def call_interface(user_id: int) -> JSONStr:
     """
     Генерирует клавиатуру для выбора призываемых
@@ -74,7 +29,7 @@ def call_interface(user_id: int) -> JSONStr:
     Returns:
         JSONStr: Строка с клавиатурой
     """
-    kb = alphabet(user_id)
+    kb = kbs.common.alphabet(user_id)
     if len(kb.buttons[-1]):
         kb.add_row()
     kb.add_text_button(text="✅ Сохранить", payload={"button": "save_selected"})
@@ -121,18 +76,6 @@ def list_of_students(letter: str, user_id: int) -> JSONStr:
     return kb.get_keyboard()
 
 
-def prompt() -> Keyboard:
-    """
-    Генерирует клавиатуру с подтверждением действия
-    Returns:
-        Keyboard: Объект клавиатуры
-    """
-    kb = Keyboard()
-    kb.add_text_button(text="✅ Подтвердить", payload={"button": "confirm"})
-    kb.add_text_button(text="🚫 Отменить", payload={"button": "deny"})
-    return kb
-
-
 def call_prompt(admin_id: int) -> JSONStr:
     """
     Генерирует клавиатуру с подтверждением отправки призыва и возможностью его
@@ -144,7 +87,7 @@ def call_prompt(admin_id: int) -> JSONStr:
     Returns:
         JSONStr:  Клавиатура
     """
-    kb = prompt()
+    kb = kbs.common.prompt()
     kb.add_row()
     store = db.admin.get_admin_storage(admin_id)
     if store.names_usage:
